pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "video_archive"
version = "0.1.0"
description = "Simple app for maintaining your video archive"
authors = ["nikita <davydovs59@gmail.ru>"]

[tool.poetry.dependencies]
python = "3.9.2"
Django = "3.1.7"
uwsgi = "2.0.19.1"
mysqlclient = "2.0.3"
django-split-settings = "1.0"
django-health-check = "3.16"
django-model-utils = "4.0.0"
django-video-encoding = "1.0.0"
structlog = "21.1"
python-decouple = "3.4"
Pillow = "8.1.2"
Celery = "4.4.7"
flower = "0.9.5"
django-celery-results = "2.0.1"
django-crispy-forms = "1.11.2"
django-bootstrap-pagination = "1.7.1"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "3.2"
django-coverage-plugin = "1.8"
wemake-python-styleguide = "^0.15"
flake8-pytest-style = "^1.4"
flake8-django = "^1.1"
flake8-logging-format = "^0.6"

pytest = "^6.2"
pytest-django = "^4.1"
pytest-cov = "^2.11"
pytest-randomly = "^3.4"
pytest-deadfixtures = "^2.2"
pytest-testmon = "^1.1"
pytest-timeout = "^1.4"
django-test-migrations = "^1.1"
hypothesis = "^6.8"

dotenv-linter = "^0.2"
